refactor(PA001): move validation logic to rules/pa_series.py
Move PluginAgentFrontmatterValidator logic out of the monolith into
rules/pa_series.py where @skilllint_rule decorates the real validation
function. Rule details for `skilllint rule PA001` come from the
function's docstring — no duplication.

The validator pipeline still needs a Validator protocol class, so
pa_series.py exports a thin adapter that delegates to check_pa001().
This is a documented architecture gap for future cleanup.
